+++ capi/example/examples/gain_stm.h
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 30/05/2022
+// Last Modified: 10/06/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -24,12 +24,12 @@
 void gain_stm(void* autd) {
   void* s = NULL;
   AUTDCreateSilencer(&s, 0xFFFF, 4096);
-  AUTDSendHeader(autd, s);
+  AUTDSend(autd, s, NULL);
   AUTDDeleteSilencer(s);
 
   void* m = NULL;
   AUTDModulationStatic(&m, 0xFF);
-  AUTDSendHeader(autd, m);
+  AUTDSend(autd, m, NULL);
 
   double x = TRANS_SPACING_MM * (((double)NUM_TRANS_X - 1.0) / 2.0);
   double y = TRANS_SPACING_MM * (((double)NUM_TRANS_Y - 1.0) / 2.0);
@@ -49,12 +49,12 @@ void gain_stm(void* autd) {
   }
 
   const uint32_t v = AUTDSTMSamplingFrequencyDivision(stm);
-  printf_s("Sample frequency division is %d\n", v);
+  printf("Sample frequency division is %d\n", v);
 
   const double actual_freq = AUTDSTMSetFrequency(stm, 1.0);
-  printf_s("Actual frequency is %lf Hz\n", actual_freq);
+  printf("Actual frequency is %lf Hz\n", actual_freq);
 
-  AUTDSendBody(autd, stm);
+  AUTDSend(autd, NULL, stm);
 
   AUTDDeleteSTM(stm);
   AUTDDeleteModulation(m);