+++ include/autd3/driver/cpu/header.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 30/05/2022
+// Last Modified: 10/06/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -17,12 +17,6 @@
 #include "defined.hpp"
 
 namespace autd3::driver {
-
-struct SyncHeader {
-  uint16_t ecat_sync_cycle_ticks;
-  uint16_t pad;
-  uint8_t data[120];
-};
 
 struct ModHead {
   uint32_t freq_div;
@@ -53,9 +47,6 @@ struct GlobalHeader {
 
   [[nodiscard]] const ModBody& mod_body() const noexcept { return *reinterpret_cast<ModBody const*>(&data[0]); }
   ModBody& mod_body() noexcept { return *reinterpret_cast<ModBody*>(&data[0]); }
-
-  [[nodiscard]] const SyncHeader& sync_header() const noexcept { return *reinterpret_cast<SyncHeader const*>(&data[0]); }
-  SyncHeader& sync_header() noexcept { return *reinterpret_cast<SyncHeader*>(&data[0]); }
 
   [[nodiscard]] const SilencerHeader& silencer_header() const noexcept { return *reinterpret_cast<SilencerHeader const*>(&data[0]); }
   SilencerHeader& silencer_header() noexcept { return *reinterpret_cast<SilencerHeader*>(&data[0]); }