+++ tools/wrapper_generator/src/main.rs
@@ -4,7 +4,7 @@
  * Created Date: 24/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 30/05/2022
+ * Last Modified: 07/06/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -12,6 +12,8 @@
  */
 
 mod csharp;
+mod matlab;
+mod nim;
 mod python;
 mod traits;
 
@@ -21,6 +23,8 @@ use anyhow::Result;
 
 use capi_header_parser::{cmake, parse::parse};
 use csharp::CSharpGenerator;
+use matlab::MatlabGenerator;
+use nim::NimGenerator;
 use python::PythonGenerator;
 use traits::Generator;
 
@@ -44,11 +48,8 @@ fn gen<G: Generator, P: AsRef<Path>>(path: P, capi_path: P) -> Result<()> {
 fn main() -> Result<()> {
     gen::<PythonGenerator, _>("python", "../../capi")?;
     gen::<CSharpGenerator, _>("cs", "../../capi")?;
-
-    // let functions = parse::parse(proj.header())?;
-    // for func in functions {
-    //     dbg!(func);
-    // }
+    gen::<MatlabGenerator, _>("matlab", "../../capi")?;
+    gen::<NimGenerator, _>("nim", "../../capi")?;
 
     Ok(())
 }